Rename buy postage to postage for simpler overall naming
